pyproject.toml
[project]
name = "ip-crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "colorama>=0.4.6",
    "crewai>=1.7.1",
    "crewai-tools>=1.7.1",
    "fastapi>=0.124.4",
    "langchain-openai>=0.3.23",
    "pandas>=2.3.3",
    "pyfiglet>=1.0.4",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "reportlab>=4.4.6",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "streamlite",
    "utils>=1.0.2",
]
